Update and rename MaxColour.java to Question_7.java
